Extract it into a hook
